Move common functions to global_test.go
